+++ NEWS.md
@@ -5,6 +5,10 @@
 * `ggpredict()` now works for models of class `clm2` from package *ordinal*,
   however, confidence intervals are not yet supported for these models.
 
+* `ggeffect()` now passes the `latent` argument for models with ordinal outcome
+  down to `effects::Effect()`, to plot effects for ordinal models on the latent
+  scale.
+
 * When argument `test` in `test_predictions()` is `"interaction"`,
   `"consecutive"`, or a data frame, *emmeans* is automatically used as backend,
   as this is the relevant package that supports these argument types.
@@ -31,6 +35,11 @@
 * Several arguments have been deprecated and replaced by new argument names. A
   warning is printed when deprecated arguments are used. The deprecated arguments
   will be removed in a future release.
+
+## Bug fixes
+
+* Fixed issue in `print()` for `ggeffect()` and models with ordinal outcome,
+  where one column was too much in the output.
 
 # ggeffects 1.5.2